dataExplorer_Insights/backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/LLM Libraries
langchain==0.1.4
langchain-openai==0.0.5
langchain-anthropic==0.0.2
langgraph==0.0.19
openai==1.10.0

# Database Drivers
sqlalchemy==2.0.25
psycopg2-binary>=2.9.0
pymysql==1.1.0
pymssql>=2.2.0
pymongo==4.6.1

# Cloud Data Warehouses
snowflake-connector-python==3.6.0
snowflake-sqlalchemy==1.5.1
redshift-connector==2.0.918
google-cloud-bigquery==3.14.1
google-cloud-bigquery-storage==2.24.0

# Enterprise Databases
cx-Oracle==8.3.0
oracledb==2.0.1
ibm-db==3.2.3
ibm-db-sa==0.4.0

# NoSQL Databases
cassandra-driver==3.28.0
boto3==1.34.20  # For DynamoDB
aioboto3==12.3.0  # Async DynamoDB

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Caching and Session
redis==5.0.1
hiredis==2.3.2

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Monitoring and Logging
python-json-logger==2.0.7

# Export functionality
openpyxl==3.1.2
xlsxwriter==3.1.9

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
flake8==7.0.0
mypy==1.8.0
